refactor(webapp): extract shared TaskRunsList for both task pages
